Make CalculateValue have a generic instead of always returning double + converting a bunch of doubles to floats

+++ common/src/main/java/lol/zanspace/unloadedactivity/OccurrencesAndDuration.java
@@ -3,12 +3,12 @@ package lol.zanspace.unloadedactivity;
 import com.mojang.datafixers.util.Pair;
 import net.minecraft.util.RandomSource;
 
-public record OccurrencesAndDuration (int occurrences, long duration, double averageProbability) {
+public record OccurrencesAndDuration (int occurrences, long duration, float averageProbability) {
     public static OccurrencesAndDuration empty() {
-        return new OccurrencesAndDuration(0, 0, 0.0);
+        return new OccurrencesAndDuration(0, 0, 0F);
     }
 
-    public static OccurrencesAndDuration recalculatedDuration(int occurrences, long cycles, double odds, RandomSource random) {
+    public static OccurrencesAndDuration recalculatedDuration(int occurrences, long cycles, float odds, RandomSource random) {
         return new OccurrencesAndDuration(occurrences, Utils.sampleNegativeBinomialWithMax(cycles, occurrences, odds, random), odds);
     }
 }
