+++ common/src/main/java/lol/zanspace/unloadedactivity/mixin/chunk/randomTicks/BambooSaplingMixin.java
@@ -29,7 +29,7 @@ public abstract class BambooSaplingMixin extends Block {
     @Shadow protected void growBamboo(Level level, BlockPos pos) {}
 
     @Override
-    public @Nullable Triple<BlockState, OccurrencesAndDuration, BlockPos> simulateProperty(BlockState state, ServerLevel level, BlockPos pos, SimulateProperty simulateProperty, RandomSource random, long timePassed, double randomPickOdds, boolean calculateDuration, @Nullable ActiveGroupSimulateData groupSimulateData) {
+    public @Nullable Triple<BlockState, OccurrencesAndDuration, BlockPos> simulateProperty(BlockState state, ServerLevel level, BlockPos pos, SimulateProperty simulateProperty, RandomSource random, long timePassed, float randomPickOdds, boolean calculateDuration, @Nullable ActiveGroupSimulateData groupSimulateData) {
         if (simulateProperty.isAction("grow_bamboo")) {
 
             int maxHeight = simulateProperty.maxHeight.orElse(15);
